LONGO function

This function allows the user to preset the variables and allows a faster analysis of data. Please use the shiny interface first to get an understanding of how this package analyzes data.

LONGOcmd(fileLocation, separator = ",", header = TRUE, commentChar = "!",
  species = "hsapiens_gene_ensembl", libraryType = "affy_primeview",
  multiProbes = "mean", windowSize = 200, stepSize = 40,
  windowStyle = "mean", filterData = TRUE, normalizeData = TRUE,
  controlColumn = 2)

Arguments

fileLocation
Address to the file to be analyzed

separator
Option to specify the separated value, default=","

header
Option to specify if there is a header in the file, default=TRUE

commentChar
Option to specify the commented character in the loaded file, default="!"

species
species name of the data based off the biomaRt database

libraryType
the gene identifier of the data based off the biomaRt database

multiProbes
Option to change the method for handling multiple probes going to the same gene, default=mean

windowSize
Option to alter the size of the windows to be created from the dataframe, default=200

stepSize
Option to alter the size of the steps used to create the windows, default=40

windowStyle
Option to alter the method used to create windows, default=mean

filterData
Option to filter data, default=TRUE

normalizeData
Option to normalize data, default=TRUE

controlColumn
column of the data used as a control for the statistical analysis

Value

Returns nothing