LONGO analysis

This is a internal function. Used by the package to take a formatted dataframe and analyzes it based on the user inputted values.

analyze(analyzeData, multiProbes, windowSize, stepSize, windowStyle, filterData,
  normalizeData, controlColumn)

Arguments

analyzeData
dataframe to be analyzed

multiProbes
Option to select what to do with multiple probes to a single gene, default=mean

windowSize
Option to alter the size of the windows to be created from the dataframe, default=200

stepSize
Option to alter the size of the steps used to create the windows, default=40

windowStyle
Option to alter the method used to create windows, default=mean

filterData
Option to filter data using edgeR::cpm, default=TRUE

normalizeData
Option to normalize data using preprocessCore::normalize.quantiles, default=TRUE

controlColumn
column of the data used as a control for the statistical analyses

Value

returns the simplified data